Module Stdune

alert unstable The API of this library is not stable and may change without notice.

include sig ... end

module Appendable_list : sig ... end

Appendable lists: concatenation takes O(1) time, conversion to a list takes O(n).

module Nonempty_list : sig ... end

A (to be expanded) collection of functions for working with non-empty lists.

module Ansi_color : sig ... end

module Array : sig ... end

module Bytes : sig ... end

module Char : sig ... end

module Comparator : sig ... end

module Either : sig ... end

Left or right

module Exn : sig ... end

Exceptions

module Exn_with_backtrace : sig ... end

An exception together with the backtrace that raised it.

module Filename : sig ... end

Represent a path component.

module Filename_set : sig ... end

Like Path.Set.t but tailored for representing sets of file names in the same parent directory. Compared to Path.Set.t, Filename_set.t statically enforces an important invariant, and can also be processed more efficiently.

module Format : sig ... end

module Hashtbl : sig ... end

module Table : sig ... end

Hashtable with a simple polymorphic type, but without the polymorphic equality.

module Int : sig ... end

module Id : sig ... end

module Io : sig ... end

IO operations.

module Lazy : sig ... end

module List : sig ... end

module Map : sig ... end

module Option : sig ... end

Optional values

module Or_exn : sig ... end

Either a value or an exception

module Ordering = Ordering

module Pp : sig ... end

module Result : sig ... end

Result type

module Set : sig ... end

module Signal : sig ... end

Unix Signal helpers

module Comparable : sig ... end

module Comparable_intf : sig ... end

module Staged : sig ... end

module String : sig ... end

module String_builder : sig ... end

module Bool : sig ... end

module Sexp : sig ... end

S-expressions

module Path : sig ... end

Representation of paths

module Fpath : sig ... end

Functions on paths that are represented as strings

module Univ_map : sig ... end

Universal maps

module Loc : sig ... end

module Env : sig ... end

module Env_path : sig ... end

Handle the PATH environment variable.

module Proc : sig ... end

module Type_eq : sig ... end

Equality between types. See Base.Type_equal for documentation.

module Nothing : sig ... end

Uninhabited type

module Bin : sig ... end

Binaries from the PATH

module Fdecl : sig ... end

Forward declarations

module Unit : sig ... end

module Monad : sig ... end

Monad signatures

module State : sig ... end

State monad transformer.

module Monoid : sig ... end

module Float : sig ... end

module Tuple : sig ... end

module Poly : sig ... end

module Code_error : sig ... end

A programming error that should be reported upstream

module User_error : sig ... end

Error meant for humans

module User_message : sig ... end

A message for the user

module User_warning : sig ... end

Non-fatal user errors

module Lexbuf : sig ... end

Lexing buffer utilities

module Scanf : sig ... end

Safe version of Scanf from the stdlib. Does not raise parsing errors errors.

module Sys : sig ... end

module Pid : sig ... end

module Applicative : sig ... end

module type Top_closure = sig ... end

module Top_closure : sig ... end

module Seq : sig ... end

module Temp : sig ... end

Temporary file management

module Queue : sig ... end

module Caller_id : sig ... end

Who called me?

module Dune_filesystem_stubs : sig ... end

Efficient directory listing with kinds.

module Predicate : sig ... end

Predicates are functions of type 'a -> bool

module Bytes_unit : sig ... end

module Dev_null : sig ... end

Portable access /dev/null with some shared fd's to reduce the open fd count

module Platform : sig ... end

Platform detection.

module Per_item : sig ... end

Module used to represent the (per_xxx ...) forms

module Bit_set : sig ... end

A set of elements that can be represented by a single word

module type Per_item = sig ... end

module Unix_error : sig ... end

module File_kind : sig ... end

module type Applicative = sig ... end

module type Monad = sig ... end

module type Monoid = sig ... end

val reraise : exn -> _

val compare : 'a -> 'b -> [> `Use_Poly_compare ]

type ('a, 'error) result = ('a, 'error) Result.t =

| Ok of 'a

| Error of 'error

type ('a, 'b) either = ('a, 'b) Either.t =

| Left of 'a

| Right of 'b

type ordering = Ordering.t =

| Lt

| Eq

| Gt

val sprintf : ('a, unit, string) format -> 'a

val ksprintf : (string -> 'a) -> ('b, unit, string, 'a) format4 -> 'b

val printfn : ('a, unit, string, unit) format4 -> 'a

module For_tests : sig ... end