Module Stdune.Bin

Binaries from the PATH

val path_sep : char

Character used to separate entries in PATH and similar environment variables

val parse : ?sep:char -> string -> string list

Parse a PATH like variable

val parse_path : ?sep:char -> string -> Path.t list

Parse a PATH like variable expecting each element to be a path

val encode_strings : string list -> string

Return a delimited string encoding of a list of strings, such as is used by the PATH variable

val cons_path : ?path_sep:char -> Path.t -> _PATH:string option -> string

Add an entry to the contents of a PATH variable.

val exe : string

Extension to append to executable filenames

val add_exe : string -> string

Adds an .exe suffix unless it is already present

val exists : Path.t -> bool

Check if a file exists

val which : path:Path.t list -> string -> Path.t option

Look for a program in the PATH