Module Stdune.Char

include module type of struct include Char end

val code : char -> int

Return the ASCII code of the argument.

val chr : int -> char

Return the character with the given ASCII code.

raises Invalid_argument
if the argument is outside the range 0--255.

val escaped : char -> string

Return a string representing the given character, with special characters escaped following the lexical conventions of OCaml. All characters outside the ASCII printable range (32..126) are escaped, as well as backslash, double-quote, and single-quote.

val lowercase_ascii : char -> char

Convert the given character to its equivalent lowercase character, using the US-ASCII character set.

val uppercase_ascii : char -> char

Convert the given character to its equivalent uppercase character, using the US-ASCII character set.

type t = char

An alias for the type of characters.

val equal : t -> t -> bool

The equal function for chars.

val seeded_hash : int -> t -> int

A seeded hash function for characters, with the same output value as Hashtbl.seeded_hash. This function allows this module to be passed as argument to the functor Hashtbl.MakeSeeded.

val is_digit : t -> bool

Check if a character belongs to the set {'0'..'9'}.

val is_lowercase_hex : t -> bool

Check if a character belongs to the set {'0'..'9', 'a'..'f'}.

val hash : t -> int

val compare : t -> t -> Ordering.t