Module Path.Source

In the source section of the current workspace.

type w

type t = w Local_gen.t

val hash : t -> int

val to_string : t -> string

val of_string : string -> t

val parse_string_exn : loc:Stdune__.Loc0.t -> string -> t

a directory is smaller than its descendants

include Comparator.S with type t := t

val compare : t -> t -> Ordering.t

include Comparator.OPS with type t := t

val (=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val (<=) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val to_dyn : t -> Dyn.t

val extension : t -> string

val set_extension : t -> ext:string -> t

set_extension path ~ext replaces extension of path by ext

val map_extension : t -> f:(string -> string) -> t

map_extension path ~f replaces extension of path by f extension

val split_extension : t -> t * string

val basename : t -> Filename.t

val basename_opt : t -> Filename.t option

val extend_basename : t -> suffix:Filename.t -> t

module Map : Map.S with type key = t

module Set : sig ... end

val equal : t -> t -> bool

val relative : ?error_loc:Stdune__.Loc0.t -> t -> string -> t

val to_string_maybe_quoted : t -> string

val is_descendant : t -> of_:t -> bool

val is_root : t -> bool

val parent_exn : t -> t

val parent : t -> t option

val unlink_no_err : t -> unit

val root : t

module L : sig ... end

val of_local : Local.t -> t

val split_first_component : t -> (Filename.t * Local.t) option

val explode : t -> Filename.t list

val is_in_build_dir : t -> bool

Source.t does not statically forbid overlap with build directory, even though having such paths is almost always an error.

val append_local : t -> Local.t -> t

val descendant : t -> of_:t -> t option

val to_local : t -> Local.t

module Table : Hashtbl.S with type key = t