Module User_message.Style

Symbolic styles that can be used inside messages. These styles are later converted to actual concrete styles depending on the output device. For instance, when printed to the terminal they are converted to ansi terminal styles (Ansi_color.Style.t list values).

type t =

| Loc

| Error

| Warning

| Kwd

| Id

| Prompt

| Hint

| Details

| Ok

| Debug

| Success

| Ansi_styles of Ansi_color.Style.t list

val to_dyn : t -> Dyn.t

val compare : t -> t -> Ordering.t